move HealthTalismanTicker into HealthTalisman class
