+++ src/main/java/io/github/pylonmc/pylon/base/PylonBase.java
@@ -1,6 +1,6 @@
 package io.github.pylonmc.pylon.base;
 
-import io.github.pylonmc.pylon.base.items.HealthTalismanTicker;
+import io.github.pylonmc.pylon.base.items.HealthTalisman;
 import io.github.pylonmc.pylon.base.items.Sprinkler;
 import io.github.pylonmc.pylon.core.addon.PylonAddon;
 import org.bukkit.Bukkit;
@@ -23,7 +23,7 @@ public class PylonBase extends JavaPlugin implements PylonAddon {
         PylonItems.register();
         PylonBlocks.register();
         Bukkit.getPluginManager().registerEvents(new Sprinkler.SprinklerPlaceListener(), this);
-        new HealthTalismanTicker().runTaskTimer(this, 0, 40);
+        new HealthTalisman.HealthTalismanTicker().runTaskTimer(this, 0, 40);
     }
 
     @Override
